+++ src/Castle.Core/DynamicProxy/Generators/InvocationTypeGenerator.cs
@@ -299,18 +299,24 @@ namespace Castle.DynamicProxy.Generators
 
 		private void ImplementChangeProxyTarget(AbstractTypeEmitter invocation, ClassEmitter @class)
 		{
-			var changeInvocationTarget = invocation.CreateMethod("ChangeProxyTarget", typeof(void), new[] { typeof(object) });
-			changeInvocationTarget.CodeBuilder.AddStatement(
+			var changeProxyTarget = invocation.CreateMethod("ChangeProxyTarget", typeof(void), new[] { typeof(object) });
+
+			var proxyObject = new FieldReference(InvocationMethods.ProxyObject);
+			var localProxy = changeProxyTarget.CodeBuilder.DeclareLocal(typeof(IProxyTargetAccessor));
+			changeProxyTarget.CodeBuilder.AddStatement(
+				new AssignStatement(localProxy,
+					new ConvertExpression(localProxy.Type, proxyObject.ToExpression())));
+
+			var dynSetProxy = typeof(IProxyTargetAccessor).GetMethod(nameof(IProxyTargetAccessor.DynProxySetTarget));
+
+			changeProxyTarget.CodeBuilder.AddStatement(
 				new ExpressionStatement(
-					new ConvertExpression(@class.TypeBuilder.AsType(), new FieldReference(InvocationMethods.ProxyObject).ToExpression())));
+					new MethodInvocationExpression(localProxy, dynSetProxy, changeProxyTarget.Arguments[0].ToExpression())
+					{
+						VirtualCall = true
+					}));
 
-			var field = @class.GetField("__target");
-			changeInvocationTarget.CodeBuilder.AddStatement(
-				new AssignStatement(
-					new FieldReference(field.Reference) { OwnerReference = null },
-					new ConvertExpression(field.Fieldbuilder.FieldType, changeInvocationTarget.Arguments[0].ToExpression())));
-
-			changeInvocationTarget.CodeBuilder.AddStatement(new ReturnStatement());
+			changeProxyTarget.CodeBuilder.AddStatement(new ReturnStatement());
 		}
 
 		private void ImplementChangeProxyTargetInterface(ClassEmitter @class, AbstractTypeEmitter invocation,